+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Terra Verbose Reporter that logs jest test results to a json file.
+
 ### Removed (passively)
 * removed dockerfiles and the travis step to release dockerfiles. Moved to a new repo: https://github.com/cerner/terra-docker